add excel streaming to $everything
