+++ src/operations/fhirOperationsManager.js
@@ -1,41 +1,46 @@
-const { SearchBundleOperation } = require('./search/searchBundle');
-const { SearchByIdOperation } = require('./searchById/searchById');
-const { ExpandOperation } = require('./expand/expand');
-const { ExportOperation } = require('./export/export');
-const { CreateOperation } = require('./create/create');
-const { UpdateOperation } = require('./update/update');
-const { MergeOperation } = require('./merge/merge');
-const { EverythingOperation } = require('./everything/everything');
-const { RemoveOperation } = require('./remove/remove');
-const { SearchByVersionIdOperation } = require('./searchByVersionId/searchByVersionId');
-const { HistoryOperation } = require('./history/history');
-const { HistoryByIdOperation } = require('./historyById/historyById');
-const { PatchOperation } = require('./patch/patch');
-const { ValidateOperation } = require('./validate/validate');
-const { GraphOperation } = require('./graph/graph');
-const { get_all_args } = require('./common/get_all_args');
-const { FhirRequestInfo } = require('../utils/fhirRequestInfo');
-const { SearchStreamingOperation } = require('./search/searchStreaming');
-const { assertTypeEquals, assertIsValid } = require('../utils/assertType');
+const {SearchBundleOperation} = require('./search/searchBundle');
+const {SearchByIdOperation} = require('./searchById/searchById');
+const {ExpandOperation} = require('./expand/expand');
+const {ExportOperation} = require('./export/export');
+const {CreateOperation} = require('./create/create');
+const {UpdateOperation} = require('./update/update');
+const {MergeOperation} = require('./merge/merge');
+const {EverythingOperation} = require('./everything/everything');
+const {RemoveOperation} = require('./remove/remove');
+const {SearchByVersionIdOperation} = require('./searchByVersionId/searchByVersionId');
+const {HistoryOperation} = require('./history/history');
+const {HistoryByIdOperation} = require('./historyById/historyById');
+const {PatchOperation} = require('./patch/patch');
+const {ValidateOperation} = require('./validate/validate');
+const {GraphOperation} = require('./graph/graph');
+const {get_all_args} = require('./common/get_all_args');
+const {FhirRequestInfo} = require('../utils/fhirRequestInfo');
+const {SearchStreamingOperation} = require('./search/searchStreaming');
+const {assertTypeEquals, assertIsValid} = require('../utils/assertType');
 const httpContext = require('express-http-context');
-const { FhirResponseStreamer } = require('../utils/fhirResponseStreamer');
+const {FhirResponseStreamer} = require('../utils/fhirResponseStreamer');
 const BundleEntry = require('../fhir/classes/4_0_0/backbone_elements/bundleEntry');
-const { convertErrorToOperationOutcome } = require('../utils/convertErrorToOperationOutcome');
+const {convertErrorToOperationOutcome} = require('../utils/convertErrorToOperationOutcome');
 const contentType = require('content-type');
-const { QueryRewriterManager } = require('../queryRewriters/queryRewriterManager');
-const { R4ArgsParser } = require('./query/r4ArgsParser');
-const { REQUEST_ID_TYPE } = require('../constants');
-const { shouldStreamResponse } = require('../utils/requestHelpers');
-const { ParametersBodyParser } = require('./common/parametersBodyParser');
-const { fhirContentTypes, hasNdJsonContentType, hasCsvContentType, hasExcelContentType} = require('../utils/contentTypes');
-const { ExportByIdOperation } = require('./export/exportById');
-const { FhirResponseNdJsonStreamer } = require('../utils/fhirResponseNdJsonStreamer');
-const { READ, WRITE } = require('../constants').OPERATIONS;
+const {QueryRewriterManager} = require('../queryRewriters/queryRewriterManager');
+const {R4ArgsParser} = require('./query/r4ArgsParser');
+const {REQUEST_ID_TYPE} = require('../constants');
+const {shouldStreamResponse} = require('../utils/requestHelpers');
+const {ParametersBodyParser} = require('./common/parametersBodyParser');
+const {
+    fhirContentTypes,
+    hasNdJsonContentType,
+    hasCsvContentType,
+    hasExcelContentType
+} = require('../utils/contentTypes');
+const {ExportByIdOperation} = require('./export/exportById');
+const {FhirResponseNdJsonStreamer} = require('../utils/fhirResponseNdJsonStreamer');
+const {READ, WRITE} = require('../constants').OPERATIONS;
 const accepts = require("accepts");
-const { vulcanIgSearchQueries } = require('./query/customQueries');
-const { ParsedArgs } = require('./query/parsedArgs');
-const { getNestedValueByPath } = require('../utils/object');
-const { ConfigManager } = require('../utils/configManager');
+const {vulcanIgSearchQueries} = require('./query/customQueries');
+const {ParsedArgs} = require('./query/parsedArgs');
+const {getNestedValueByPath} = require('../utils/object');
+const {ConfigManager} = require('../utils/configManager');
 const {FhirResponseCsvStreamer} = require("../utils/fhirResponseCsvStreamer");
 const {FhirResponseExcelStreamer} = require("../utils/fhirResponseExcelStreamer");
 const {SummaryOperation} = require("./summary/summary");
@@ -207,7 +212,7 @@ class FhirOperationsManager {
      * @param {import('http').IncomingMessage} req
      * @return {FhirRequestInfo}
      */
-    getRequestInfo (req) {
+    getRequestInfo(req) {
         assertIsValid(req, 'req is null');
         /**
          * @type {string | null}
@@ -246,7 +251,7 @@ class FhirOperationsManager {
         /**
          * @type {string|string[]|null}
          */
-        const accept =  accepts(req).types();
+        const accept = accepts(req).types();
         /**
          * @type {string}
          */
@@ -308,12 +313,12 @@ class FhirOperationsManager {
      * @param {string} operation
      * @return {Promise<ParsedArgs>}
      */
-    async getParsedArgsAsync ({ args, resourceType, headers, operation }) {
-        const { base_version } = args;
+    async getParsedArgsAsync({args, resourceType, headers, operation}) {
+        const {base_version} = args;
         /**
          * @type {ParsedArgs}
          */
-        let parsedArgs = this.r4ArgsParser.parseArgs({ resourceType, args });
+        let parsedArgs = this.r4ArgsParser.parseArgs({resourceType, args});
         // see if any query rewriters want to rewrite the args
         parsedArgs = await this.queryRewriterManager.rewriteArgsAsync(
             {
@@ -333,13 +338,13 @@ class FhirOperationsManager {
      * @param {string} resourceType
      * @return {Resource[] | Resource} array of resources
      */
-    async search (args, { req }, resourceType) {
+    async search(args, {req}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
         /**
          * @type {ParsedArgs}
          */
@@ -364,14 +369,14 @@ class FhirOperationsManager {
      * @param {string} resourceType
      * @return {Resource[] | Resource} array of resources
      */
-    async searchStreaming (args, { req, res }, resourceType) {
+    async searchStreaming(args, {req, res}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
-        const requestInfo = this.getRequestInfo(req)
+        combined_args = this.parseParametersFromBody({req, combined_args});
+        const requestInfo = this.getRequestInfo(req);
 
         /**
          * @type {ParsedArgs}
@@ -397,8 +402,8 @@ class FhirOperationsManager {
 
                         // for missing modifier handling needs to be done while applying filter to parent
                         if (
-                            (parsedArg.modifiers.length == 0 ||
-                                !parsedArg.modifiers[0] === 'missing') &&
+                            (parsedArg.modifiers.length === 0 ||
+                                parsedArg.modifiers[0] !== 'missing') &&
                             vulcanIgFilter.searchParam
                         ) {
                             vulcanIgFilterArgs[vulcanIgFilter.searchParam] =
@@ -441,9 +446,8 @@ class FhirOperationsManager {
                     const existingParseArgItem = parsedArgs.parsedArgItems.find(
                         (a) =>
                             a.queryParameter ===
-                                (vulcanIgSearchQuery.resultSearchParam === 'id'
-                                    ? '_id'
-                                    : vulcanIgSearchQuery.resultSearchParam) &&
+                            (vulcanIgSearchQuery.resultSearchParam === 'id' ? '_id'
+                                : vulcanIgSearchQuery.resultSearchParam) &&
                             a.modifiers.length === 0
                     );
                     if (existingParseArgItem) {
@@ -462,16 +466,14 @@ class FhirOperationsManager {
                     const parentResourceArgName =
                         vulcanIgSearchQuery.resultSearchParam +
                         (parsedArg.modifiers[0] === 'missing' &&
-                        parsedArg.queryParameterValue.value === 'true'
-                            ? ':not'
+                        parsedArg.queryParameterValue.value === 'true' ? ':not'
                             : '');
 
                     const parentResourceParsedArg = await this.getParsedArgsAsync({
                         args: {
                             base_version: '4_0_0',
                             [parentResourceArgName]:
-                                parentResourceFilterValue.length > 0
-                                    ? parentResourceFilterValue.join(',')
+                                parentResourceFilterValue.length > 0 ? parentResourceFilterValue.join(',')
                                     : '__invalid__' // for handling case when no result is found
                         },
                         resourceType,
@@ -492,7 +494,7 @@ class FhirOperationsManager {
             });
     }
 
-    parseParametersFromBody ({ req, combined_args }) {
+    parseParametersFromBody({req, combined_args}) {
         let args = combined_args;
         if (req.body) {
             if (req.headers['content-type'] === fhirContentTypes.form_urlencoded) {
@@ -520,13 +522,13 @@ class FhirOperationsManager {
      * @param {import('http').IncomingMessage} req
      * @param {string} resourceType
      */
-    async searchById (args, { req }, resourceType) {
+    async searchById(args, {req}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
         /**
          * @type {ParsedArgs}
          */
@@ -550,13 +552,13 @@ class FhirOperationsManager {
      * @param {string} resourceType
      * @returns {Resource}
      */
-    async create (args, { req }, resourceType) {
+    async create(args, {req}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
         /**
          * @type {string}
          */
@@ -585,13 +587,13 @@ class FhirOperationsManager {
      * @param {string} resourceType
      * @returns {Promise<{id: string,created: boolean, resource_version: string, resource: Resource}>}
      */
-    async update (args, { req }, resourceType) {
+    async update(args, {req}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
         /**
          * @type {ParsedArgs}
          */
@@ -617,10 +619,10 @@ class FhirOperationsManager {
      * @param {string} resourceType
      * @return {Promise<Resource | Resource[] | void>}
      */
-    async merge(args, { req, res }, resourceType) {
+    async merge(args, {req, res}, resourceType) {
         // Combine args
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
 
         const parsedArgs = await this.getParsedArgsAsync({
             args: combined_args,
@@ -652,7 +654,6 @@ class FhirOperationsManager {
     }
 
 
-
     /**
      * does a FHIR $everything
      * @param {string[]} args
@@ -661,7 +662,7 @@ class FhirOperationsManager {
      * @param {string} resourceType
      * @returns {Bundle}
      */
-    async everything (args, { req, res }, resourceType) {
+    async everything(args, {req, res}, resourceType) {
         /**
          * combined args
          * @type {Object}
@@ -681,15 +682,36 @@ class FhirOperationsManager {
          */
         const requestInfo = this.getRequestInfo(req);
         if (shouldStreamResponse(req)) {
-            const responseStreamer = hasNdJsonContentType(requestInfo.accept) ?
-                new FhirResponseNdJsonStreamer({
+            /**
+             * response streamer to use
+             * @type {BaseResponseStreamer}
+             */
+            let responseStreamer;
+            if (hasNdJsonContentType(requestInfo.accept)) {
+                responseStreamer = new FhirResponseNdJsonStreamer({
                         response: res,
                         requestId: req.id
                     }
-                ) : new FhirResponseStreamer({
+                );
+            } else if (hasCsvContentType(requestInfo.accept) || hasCsvContentType(parsedArgs._format)) {
+                responseStreamer = new FhirResponseCsvStreamer({
+                        response: res,
+                        requestId: req.id
+                    }
+                );
+            } else if (hasExcelContentType(requestInfo.accept) || hasExcelContentType(parsedArgs._format)) {
+                responseStreamer = new FhirResponseExcelStreamer({
+                        response: res,
+                        requestId: req.id
+                    }
+                );
+            } else {
+                responseStreamer = new FhirResponseStreamer({
                     response: res,
                     requestId: req.id
                 });
+            }
+            // Start the response streamer
             await responseStreamer.startAsync();
 
             try {
@@ -711,7 +733,7 @@ class FhirOperationsManager {
                 /**
                  * @type {OperationOutcome}
                  */
-                const operationOutcome = convertErrorToOperationOutcome({ error: err });
+                const operationOutcome = convertErrorToOperationOutcome({error: err});
                 await responseStreamer.writeBundleEntryAsync({
                         bundleEntry: new BundleEntry({
                                 resource: operationOutcome
@@ -719,7 +741,7 @@ class FhirOperationsManager {
                         )
                     }
                 );
-                await responseStreamer.setStatusCodeAsync({ statusCode: status });
+                await responseStreamer.setStatusCodeAsync({statusCode: status});
                 await responseStreamer.endAsync();
             }
         } else {
@@ -778,13 +800,13 @@ class FhirOperationsManager {
                     response: res,
                     requestId: req.id
                 }
-            )
+            );
         } else if (hasExcelContentType(requestInfo.accept) || hasExcelContentType(parsedArgs._format)) {
             responseStreamer = new FhirResponseExcelStreamer({
                     response: res,
                     requestId: req.id
                 }
-            )
+            );
         }
 
         await responseStreamer.startAsync();
@@ -827,13 +849,13 @@ class FhirOperationsManager {
      * @param {import('http').IncomingMessage} req
      * @param {string} resourceType
      */
-    async remove (args, { req }, resourceType) {
+    async remove(args, {req}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
         /**
          * @type {ParsedArgs}
          */
@@ -856,13 +878,13 @@ class FhirOperationsManager {
      * @param {import('http').IncomingMessage} req
      * @param {string} resourceType
      */
-    async remove_by_query (args, { req }, resourceType) {
+    async remove_by_query(args, {req}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
         /**
          * @type {ParsedArgs}
          */
@@ -885,13 +907,13 @@ class FhirOperationsManager {
      * @param {import('http').IncomingMessage} req
      * @param {string} resourceType
      */
-    async searchByVersionId (args, { req }, resourceType) {
+    async searchByVersionId(args, {req}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
 
         /**
          * @type {ParsedArgs}
@@ -915,13 +937,13 @@ class FhirOperationsManager {
      * @param {import('http').IncomingMessage} req
      * @param {string} resourceType
      */
-    async history (args, { req }, resourceType) {
+    async history(args, {req}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
 
         /**
          * @type {ParsedArgs}
@@ -946,13 +968,13 @@ class FhirOperationsManager {
      * @param {import('http').IncomingMessage} req
      * @param {string} resourceType
      */
-    async historyById (args, { req }, resourceType) {
+    async historyById(args, {req}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
 
         /**
          * @type {ParsedArgs}
@@ -976,13 +998,13 @@ class FhirOperationsManager {
      * @param {string} resourceType
      * @return {{id: string,created: boolean, resource_version: string, resource: Resource}}
      */
-    async patch (args, { req }, resourceType) {
+    async patch(args, {req}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
 
         /**
          * @type {ParsedArgs}
@@ -1005,13 +1027,13 @@ class FhirOperationsManager {
      * @param {import('http').IncomingMessage} req
      * @param {string} resourceType
      */
-    async validate (args, { req }, resourceType) {
+    async validate(args, {req}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
 
         /**
          * @type {ParsedArgs}
@@ -1036,7 +1058,7 @@ class FhirOperationsManager {
      * @param {string} resourceType
      * @return {Promise<Bundle>}
      */
-    async graph (args, { req, res }, resourceType) {
+    async graph(args, {req, res}, resourceType) {
         /**
          * combined args
          * @type {Object}
@@ -1088,7 +1110,7 @@ class FhirOperationsManager {
             /**
              * @type {OperationOutcome}
              */
-            const operationOutcome = convertErrorToOperationOutcome({ error: err });
+            const operationOutcome = convertErrorToOperationOutcome({error: err});
             await responseStreamer.writeBundleEntryAsync({
                     bundleEntry: new BundleEntry({
                             resource: operationOutcome
@@ -1096,7 +1118,7 @@ class FhirOperationsManager {
                     )
                 }
             );
-            await responseStreamer.setStatusCodeAsync({ statusCode: status });
+            await responseStreamer.setStatusCodeAsync({statusCode: status});
             await responseStreamer.endAsync();
         }
     }
@@ -1107,13 +1129,13 @@ class FhirOperationsManager {
      * @param {import('http').IncomingMessage} req
      * @param {string} resourceType
      */
-    async expand (args, { req }, resourceType) {
+    async expand(args, {req}, resourceType) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
         /**
          * @type {ParsedArgs}
          */
@@ -1137,19 +1159,19 @@ class FhirOperationsManager {
      * @param {{ req: import('http').IncomingMessage }}
      * @return {Resource | Resource[]}
      */
-    async export (args, { req }) {
+    async export(args, {req}) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
         /**
          * @type {FhirRequestInfo}
          */
         const requestInfo = this.getRequestInfo(req);
 
-        return await this.exportOperation.exportAsync({ requestInfo, args: combined_args });
+        return await this.exportOperation.exportAsync({requestInfo, args: combined_args});
     }
 
     /**
@@ -1158,19 +1180,19 @@ class FhirOperationsManager {
      * @param {{ req: import('http').IncomingMessage }}
      * @return {Resource | Resource[]}
      */
-    async exportById (args, { req }) {
+    async exportById(args, {req}) {
         /**
          * combined args
          * @type {Object}
          */
         let combined_args = get_all_args(req, args);
-        combined_args = this.parseParametersFromBody({ req, combined_args });
+        combined_args = this.parseParametersFromBody({req, combined_args});
         /**
          * @type {FhirRequestInfo}
          */
         const requestInfo = this.getRequestInfo(req);
 
-        return await this.exportByIdOperation.exportByIdAsync({ requestInfo, args: combined_args });
+        return await this.exportByIdOperation.exportByIdAsync({requestInfo, args: combined_args});
     }
 }
